restructing the decode portion of the software.
