+++ logs/TEST-LOG.txt
@@ -485,3 +485,5 @@
 [Thu_Nov__3_01:54:25_2022][/home/akiel/Desktop/Trunk/github/public-repos/OSS/vpaSTRM/src/vpaSTRM.c][line: 50] <-------------------------NEW-SESSION------------------------->
 [Thu_Nov__3_01:54:34_2022][/home/akiel/Desktop/Trunk/github/public-repos/OSS/vpaSTRM/src/vpaSTRM.c][line: 50] <-------------------------NEW-SESSION------------------------->
 [Thu_Nov__3_01:56:51_2022][/home/akiel/Desktop/Trunk/github/public-repos/OSS/vpaSTRM/src/vpaSTRM.c][line: 50] <-------------------------NEW-SESSION------------------------->
+[Thu_Nov__3_02:31:20_2022][/home/akiel/Desktop/Trunk/github/public-repos/OSS/vpaSTRM/src/vpaSTRM.c][line: 50] <-------------------------NEW-SESSION------------------------->
+[Thu_Nov__3_02:31:25_2022][/home/akiel/Desktop/Trunk/github/public-repos/OSS/vpaSTRM/src/vpaSTRM.c][line: 50] <-------------------------NEW-SESSION------------------------->
